tests(Testcafe_devextreme): fix imports after a11y tests moving
