+++ e2e/testcafe-devextreme/tests/accessibility/common/accordion.ts
@@ -1,11 +1,11 @@
 import { Properties } from 'devextreme/ui/accordion.d';
 import { Selector as $ } from 'testcafe';
-import url from '../../helpers/getPageUrl';
-import { testAccessibility, Configuration } from '../../helpers/accessibility/test';
-import { Options } from '../../helpers/generateOptionMatrix';
+import url from '../../../helpers/getPageUrl';
+import { testAccessibility, Configuration } from '../../../helpers/accessibility/test';
+import { Options } from '../../../helpers/generateOptionMatrix';
 
 fixture.disablePageReloads`Accessibility`
-  .page(url(__dirname, '../container.html'));
+  .page(url(__dirname, '../../container.html'));
 
 const items = ['Item_1', 'Item_2', 'Item_3'];
