+++ e2e/testcafe-devextreme/tests/accessibility/common/dateBox.ts
@@ -1,12 +1,12 @@
 import { ClientFunction } from 'testcafe';
 import { Properties } from 'devextreme/ui/date_box.d';
 import DateBox from 'devextreme-testcafe-models/dateBox';
-import url from '../../helpers/getPageUrl';
-import { defaultSelector, testAccessibility, Configuration } from '../../helpers/accessibility/test';
-import { Options } from '../../helpers/generateOptionMatrix';
+import url from '../../../helpers/getPageUrl';
+import { defaultSelector, testAccessibility, Configuration } from '../../../helpers/accessibility/test';
+import { Options } from '../../../helpers/generateOptionMatrix';
 
 fixture.disablePageReloads`Accessibility`
-  .page(url(__dirname, '../container.html'));
+  .page(url(__dirname, '../../container.html'));
 
 const now = new Date();
