+++ e2e/testcafe-devextreme/tests/accessibility/common/dropDownBox.ts
@@ -1,10 +1,10 @@
 import { Properties } from 'devextreme/ui/drop_down_box.d';
-import url from '../../helpers/getPageUrl';
-import { testAccessibility, Configuration } from '../../helpers/accessibility/test';
-import { Options } from '../../helpers/generateOptionMatrix';
+import url from '../../../helpers/getPageUrl';
+import { testAccessibility, Configuration } from '../../../helpers/accessibility/test';
+import { Options } from '../../../helpers/generateOptionMatrix';
 
 fixture.disablePageReloads`Accessibility`
-  .page(url(__dirname, '../container.html'));
+  .page(url(__dirname, '../../container.html'));
 
 const items = ['Download Trial For Visual Studio', 'Download Trial For All Platforms', 'Package Managers'];
