+++ e2e/testcafe-devextreme/tests/accessibility/common/map.ts
@@ -1,11 +1,11 @@
 import { Properties } from 'devextreme/ui/map.d';
-import url from '../../helpers/getPageUrl';
-import { testAccessibility, Configuration } from '../../helpers/accessibility/test';
-import { Options } from '../../helpers/generateOptionMatrix';
+import url from '../../../helpers/getPageUrl';
+import { testAccessibility, Configuration } from '../../../helpers/accessibility/test';
+import { Options } from '../../../helpers/generateOptionMatrix';
 
 // TODO Chrome133: skipped during chrome update
 fixture.disablePageReloads.skip`Accessibility`
-  .page(url(__dirname, '../container.html'));
+  .page(url(__dirname, '../../container.html'));
 
 const markersData = [
   { location: '40.7825, -73.966111' },
