+++ e2e/testcafe-devextreme/tests/accessibility/common/menu.ts
@@ -1,11 +1,11 @@
 import { Item, Properties } from 'devextreme/ui/menu.d';
 import Menu from 'devextreme-testcafe-models/menu';
-import url from '../../helpers/getPageUrl';
-import { testAccessibility, Configuration } from '../../helpers/accessibility/test';
-import { Options } from '../../helpers/generateOptionMatrix';
+import url from '../../../helpers/getPageUrl';
+import { testAccessibility, Configuration } from '../../../helpers/accessibility/test';
+import { Options } from '../../../helpers/generateOptionMatrix';
 
 fixture.disablePageReloads`Accessibility`
-  .page(url(__dirname, '../container.html'));
+  .page(url(__dirname, '../../container.html'));
 
 const items: Item[] = [
   {
