+++ e2e/testcafe-devextreme/tests/accessibility/common/radioGroup.ts
@@ -1,10 +1,10 @@
 import { Properties } from 'devextreme/ui/radio_group.d';
-import url from '../../helpers/getPageUrl';
-import { testAccessibility, Configuration } from '../../helpers/accessibility/test';
-import { Options } from '../../helpers/generateOptionMatrix';
+import url from '../../../helpers/getPageUrl';
+import { testAccessibility, Configuration } from '../../../helpers/accessibility/test';
+import { Options } from '../../../helpers/generateOptionMatrix';
 
 fixture.disablePageReloads`Accessibility`
-  .page(url(__dirname, '../container.html'));
+  .page(url(__dirname, '../../container.html'));
 
 const items = ['Item_1', 'Item_2', 'Item_3'];
