+++ e2e/testcafe-devextreme/tests/accessibility/common/tabPanel.ts
@@ -1,11 +1,11 @@
 import { Item, Properties } from 'devextreme/ui/tab_panel.d';
-import url from '../../helpers/getPageUrl';
-import { testAccessibility, Configuration } from '../../helpers/accessibility/test';
-import { Options } from '../../helpers/generateOptionMatrix';
-import { isMaterial, isMaterialBased } from '../../helpers/themeUtils';
+import url from '../../../helpers/getPageUrl';
+import { testAccessibility, Configuration } from '../../../helpers/accessibility/test';
+import { Options } from '../../../helpers/generateOptionMatrix';
+import { isMaterial, isMaterialBased } from '../../../helpers/themeUtils';
 
 fixture.disablePageReloads`Accessibility`
-  .page(url(__dirname, '../container.html'));
+  .page(url(__dirname, '../../container.html'));
 
 const items: Item[] = [
   { title: 'John Heart', text: 'John Heart' },
