+++ e2e/testcafe-devextreme/tests/accessibility/common/tagBox.ts
@@ -1,13 +1,13 @@
 import { Properties } from 'devextreme/ui/tag_box.d';
 import TagBox from 'devextreme-testcafe-models/tagBox';
-import url from '../../helpers/getPageUrl';
-import { defaultSelector, testAccessibility, Configuration } from '../../helpers/accessibility/test';
-import { Options } from '../../helpers/generateOptionMatrix';
+import url from '../../../helpers/getPageUrl';
+import { defaultSelector, testAccessibility, Configuration } from '../../../helpers/accessibility/test';
+import { Options } from '../../../helpers/generateOptionMatrix';
 
 const TIME_TO_WAIT = 150;
 
 fixture.disablePageReloads`Accessibility`
-  .page(url(__dirname, '../container.html'));
+  .page(url(__dirname, '../../container.html'));
 
 const items = [
   'HD Video Player',
