+++ e2e/testcafe-devextreme/tests/accessibility/common/treeView.ts
@@ -1,11 +1,11 @@
 import { Properties } from 'devextreme/ui/tree_view.d';
-import url from '../../helpers/getPageUrl';
-import { testAccessibility, Configuration } from '../../helpers/accessibility/test';
-import { Options } from '../../helpers/generateOptionMatrix';
-import { employees } from '../navigation/treeView/data';
+import url from '../../../helpers/getPageUrl';
+import { testAccessibility, Configuration } from '../../../helpers/accessibility/test';
+import { Options } from '../../../helpers/generateOptionMatrix';
+import { employees } from '../../navigation/treeView/data';
 
 fixture.disablePageReloads`Accessibility`
-  .page(url(__dirname, '../container.html'));
+  .page(url(__dirname, '../../container.html'));
 
 const options: Options<Properties> = {
   height: [
